rename ann package to neural for more generalizability
For other types of neural networks added in the future,
this expands the scope of this package.
