+++ neural/doc.go
@@ -1,0 +1,2 @@
+// Package neural provides neural network implementations.
+package neural
